requirements.txt
appdirs==1.4.3
astroid==1.6.4
backports.functools-lru-cache==1.5
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
Django==1.11
django-admin==1.2.4
django-cors-headers==1.3.1
django-excel-response2==2.0.8
django-six==1.0.4
enum34==1.1.6
futures==3.2.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
oauthlib==2.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.4.33
pycrypto==2.6.1
PyJWT==1.4.2
pylint==1.9.1
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
python-crontab==2.1.1
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
screen==1.0.1
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
wrapt==1.10.11
xlwt==1.3.0
django-filter==1.1.0
django-markdown==0.8.4
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
